fk/luae_na/src/test/java/se/inera/intyg/common/luae_na/v1/model/converter/certificate/question/QuestionGrundForMedicinsktUnderlagTest.java
/*
 * Copyright (C) 2022 Inera AB (http://www.inera.se)
 *
 * This file is part of sklintyg (https://github.com/sklintyg).
 *
 * sklintyg is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * sklintyg is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

package se.inera.intyg.common.luae_na.v1.model.converter.certificate.question;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Nested;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.TestInstance;
import org.junit.jupiter.api.TestInstance.Lifecycle;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import se.inera.intyg.common.services.texts.CertificateTextProvider;

@ExtendWith(MockitoExtension.class)
class QuestionGrundForMedicinsktUnderlagTest {

    @Mock
    private CertificateTextProvider texts;

    @BeforeEach
    void setup() {
        //when(texts.get(any(String.class))).thenReturn("Test string");
    }

    @Nested
    class ToCertificate {
        @Test
        void shouldIncludeId() {
            final var question = QuestionGrundForMedicinsktUnderlag.toCertificate( 0, texts);
            assertEquals("1.1", question.getId());
        }

        @Test
        void shouldIncludeIndex() {
            final var expectedIndex = 1;
            final var question = QuestionGrundForMedicinsktUnderlag.toCertificate( expectedIndex, texts);
            assertEquals(expectedIndex, question.getIndex());
        }

        @Test
        void shouldIncludeParentId() {
            final var question = QuestionGrundForMedicinsktUnderlag.toCertificate( 0, texts);
            assertEquals("1", question.getParent());
        }
    }
}